requirements.txt
# Django and REST Framework
django==4.2.26
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.9.0
dj-database-url==2.1.0
drf-spectacular==0.27.2

# Database
psycopg2==2.9.10

# Environment variables
python-dotenv==1.0.1

# AI Service
groq==0.11.0

# HTTP client (if needed)
httpx==0.27.2

# PDF processing
pypdf==4.0.1